pyproject.toml
[tool.poetry]
name = "robotframework-airtest"
version = "0.1.2"
description = "网易Airtest/pocoui的Robotframework测试库"
authors = ["kaluluosi <kaluluosi@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "robotframework_airtest" }]
repository = "https://github.com/kaluluosi/robotframework-airtest"

[tool.poetry.plugins."robotframework_airtest.vmg.generator"]
unity = "robotframework_airtest.vmg.generators.unity:generate"

[tool.poetry.plugins."robotframework_airtest.device.connect_strategy"]
android = "robotframework_airtest.device.connects.impl.android:AndroidConnectStrategy"
windows = "robotframework_airtest.device.connects.impl.windows:WindowsConnectStrategy"
unity = "robotframework_airtest.device.connects.impl.unity:UnityConnectStrategy"


[tool.poetry.scripts]
ra = "robotframework_airtest.cli:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.1"
black = "^23.10.0"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.4.6"
pymdown-extensions = "^10.3.1"
pygments = "^2.16.1"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "default"

[tool.black]
line-length = 88

[tool.ruff]
ignore = ['E501']

[tool.poetry.dependencies]
python = ">=3.8,<=3.9"
robotframework = "^6.1.1"
click = "^8.1.7"
pocoui = "^1.0.89"
unityparser = "^3.0.1"
pydantic-settings = "^2.0.3"
toml = "^0.10.2"
docutils = "^0.20.1"
pymdown-extensions = "^10.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
